packages/pdf_fetcher/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf_fetcher"
version = "0.1.0"
description = "Automated academic PDF downloader with publisher-specific strategies and Unpaywall integration"
readme = "README.md"
authors = [
    {name = "Henrik Kragh Sørensen", email = "hkragh@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
keywords = ["pdf", "academic", "papers", "doi", "download", "unpaywall", "open-access"]

dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "pyyaml>=6.0",
    "colorama>=0.4.6",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
selenium = [
    "selenium>=4.0.0",
]
vpn = [
    "network-utils @ file:///Users/fvb832/Documents/dh4pmp_tools/packages/network_utils",
]

[project.urls]
Repository = "ssh://hkragh@git.henrikkragh.dk/git-srv/pdf_fetcher.git"

[project.scripts]
pdf_fetcher = "pdf_fetcher.cli:main"

[tool.setuptools]
packages = ["pdf_fetcher", "pdf_fetcher.strategies"]

[tool.setuptools.package-data]
pdf_fetcher = ["config.yaml"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=pdf_fetcher --cov-report=term-missing"
